third_party/blink/web_tests/platform/generic/storage/indexeddb/keyrange-required-arguments-expected.txt

Test IndexedDB keyrange required arguments

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS IDBKeyRange.only(); threw exception TypeError: Failed to execute 'only' on 'IDBKeyRange': 1 argument required, but only 0 present..
PASS IDBKeyRange.lowerBound(); threw exception TypeError: Failed to execute 'lowerBound' on 'IDBKeyRange': 1 argument required, but only 0 present..
PASS IDBKeyRange.upperBound(); threw exception TypeError: Failed to execute 'upperBound' on 'IDBKeyRange': 1 argument required, but only 0 present..
PASS IDBKeyRange.bound(1); threw exception TypeError: Failed to execute 'bound' on 'IDBKeyRange': 2 arguments required, but only 1 present..
PASS IDBKeyRange.bound(); threw exception TypeError: Failed to execute 'bound' on 'IDBKeyRange': 2 arguments required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
